+++ release-notes.md
@@ -1,4 +1,5 @@
 `0.1.10` - ?
+* Add count for Adversary status ([#40](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/40))
 * Allow the removal of custom attack animations per item ([#39](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/39))
 * Fix for incomplete localization of dice spender (thanks `@prolice`!)
 * Add French localization (thanks `@prolice`!)